Update .d.ts files definiions
+ Adds preact so it builds correctly for the typings file.
+ Removes type roots because they are not needed.
+ Filxes Element naming and adds JSX namespace
