+++ tsconfig.json
@@ -1,7 +1,12 @@
 {
   "compileOnSave": false,
+  "files": [
+    "lib/filedrop.ts",
+    "./node_modules/preact/dist/preact"
+  ],
   "compilerOptions": {
     "strict": true,
+    "declaration": true,
     "target": "es2017",
     "module": "esnext",
     "moduleResolution": "node",
@@ -11,10 +16,6 @@
     "jsx": "react",
     "jsxFactory": "h",
     "allowJs": false,
-    "baseUrl": ".",
-    "typeRoots": [
-      "filedrop.d.ts",
-      "node_modules/@types"
-    ]
+    "baseUrl": "."
   }
 }
